Create/update data access hooks
